add ExampleDatabaseInit
